test(sqla_factory): moved the models and the fixtures to models.py and conftest.py; added a test for using coroutine_field
