+++ tests/sqlalchemy_factory/conftest.py
@@ -1,0 +1,33 @@
+from typing import AsyncIterator
+
+import pytest
+from sqlalchemy import (
+    create_engine,
+)
+from sqlalchemy.engine import Engine
+from sqlalchemy.ext.asyncio import AsyncEngine, AsyncSession, async_sessionmaker, create_async_engine
+
+from tests.sqlalchemy_factory.models import Base
+
+
+@pytest.fixture()
+def engine() -> Engine:
+    return create_engine("sqlite:///:memory:")
+
+
+@pytest.fixture()
+def async_engine() -> AsyncEngine:
+    return create_async_engine("sqlite+aiosqlite:///:memory:")
+
+
+@pytest.fixture()
+def async_session_maker(async_engine: AsyncEngine) -> async_sessionmaker[AsyncSession]:
+    return async_sessionmaker(bind=async_engine, expire_on_commit=False)
+
+
+@pytest.fixture(autouse=True)
+async def fx_drop_create_meta(async_engine: AsyncEngine) -> AsyncIterator[None]:
+    async with async_engine.begin() as conn:
+        await conn.run_sync(Base.metadata.drop_all)
+        await conn.run_sync(Base.metadata.create_all)
+    yield
